Remove api data from intial state, add middleware for initialization
